python/requirements.txt
moviepy==1.0.3
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
requests==2.31.0
python-dotenv==1.0.0